script updates

+++ tests/test_on_emolecule.py
@@ -1,56 +1,40 @@
-"""This script is specifically for testing the large eMolecules dataset.
-See tests/README.md for instructions on how to download, process, and
-test the set for developers interested in making a PR to the repository!
+"""This script is specifically for testing the large eMolecules dataset,
+which upon downloading, should have the file name version.smi.gz.
 
-This file is split into text files of 1 million SMILES strings,
-such that the developer can end the test and run from the last
-file tested, by passing the specific paths into 'datasets = []'.
+This test will automatically create a checkpoint file, so that the test
+can be paused and resumed easily. Please note that there are some SMILES
+in the dataset that are expected to fail; these are SMILES which use
+the same ring-bond number across a dot symbol.
+
+Example: C1.C2.C12, COC(=O)C1CCCN1CP123OCCO1.C(CO2)O3
 """
 
 import faulthandler
 import os
-import random
 
 import pandas as pd
 import pytest
 from rdkit.Chem import MolFromSmiles, MolToSmiles
 
 import selfies as sf
-from selfies.encoder import _parse_smiles
-from selfies.kekulize import BRANCH_TYPE, RING_TYPE, kekulize_parser
 
 faulthandler.enable()
 
 # Test Configuration ==========================================================
-# TODO: Edit the configurations of the eMolecule script below.
 
-datasets = []
+# file path
+curr_dir = os.path.dirname(__file__)
+EMOL_PATH = os.path.join(curr_dir, 'test_sets', 'version.smi.gz')
 
 # SMILES in eMolecules are under this column name
 COL_NAME = 'isosmiles'
-
-# How many samples from each data file to test. -1 to test the whole file.
-NUM_SAMPLES = -1
-
-# Dynamically search for all text files in the eMolecules folder
-curr_dir = os.path.dirname(__file__)
-emol_dir = os.path.join(curr_dir, 'test_sets', 'eMolecules')
-
-for file in os.listdir(emol_dir):
-    if file.endswith(".txt"):
-        datasets.append(file[:-4])  # remove the .txt extension
-
-
-# Alternatively, specify the files you want to test:
-# e.g. datasets = ['split05', 'split06']
 
 
 # Tests =======================================================================
 
 # TODO: Comment out the pytest skip to use this script.
 @pytest.mark.skip(reason="eMolecules dataset not on GitHub")
-@pytest.mark.parametrize("test_name", datasets)
-def test_roundtrip_translation(test_name):
+def test_roundtrip_translation():
     """Tests a roundtrip SMILES -> SELFIES -> SMILES translation of the
     SMILES examples in QM9, NonFullerene, Zinc, etc.
     """
@@ -63,30 +47,37 @@ def test_roundtrip_translation(test_name):
     sf.set_semantic_constraints(constraints)
 
     # file I/O
-    curr_dir = os.path.dirname(__file__)
-    test_path = os.path.join(curr_dir, 'test_sets', 'eMolecules',
-                             f"{test_name}.txt")
-    error_path = os.path.join(curr_dir,
-                              'error_sets',
-                              f"errors_{test_name}.csv")
+    ckpt_path = os.path.join(curr_dir, 'checkpoints', f"emolecule_ckpt.txt")
+    error_path = os.path.join(curr_dir, 'error_sets', f"errors_emolecules.csv")
 
-    os.makedirs(os.path.dirname(error_path), exist_ok=True)
+    if os.path.exists(ckpt_path):
+        with open(ckpt_path, 'r') as ckpt_file:
+            checkpoint = int(ckpt_file.readlines()[0])
+
+    else:
+        os.makedirs(os.path.dirname(ckpt_path), exist_ok=True)
+        os.makedirs(os.path.dirname(error_path), exist_ok=True)
+
+        with open(error_path, "w+") as error_log:
+            error_log.write("In, Out\n")
+        checkpoint = -1
+
     error_list = []
-    with open(error_path, "w+") as error_log:
-        error_log.write("In, Out\n")
     error_found_flag = False
 
     # make pandas reader
-    N = sum(1 for _ in open(test_path)) - 1
-    S = NUM_SAMPLES if (0 < NUM_SAMPLES <= N) else N
-    skip = sorted(random.sample(range(1, N + 1), N - S))
-    reader = pd.read_csv(test_path,
+    reader = pd.read_csv(EMOL_PATH,
                          chunksize=10000,
-                         header=0,
-                         skiprows=skip)
+                         compression='gzip',
+                         delimiter=' ',
+                         header=0)
 
     # roundtrip testing
-    for chunk in reader:
+    for chunk_idx, chunk in enumerate(reader):
+
+        if chunk_idx <= checkpoint:
+            continue
+
         for in_smiles in chunk[COL_NAME]:
 
             if (MolFromSmiles(in_smiles) is None) or ('*' in in_smiles):
@@ -109,78 +100,11 @@ def test_roundtrip_translation(test_name):
         error_found_flag = error_found_flag or error_list
         error_list = []
 
+        with open(ckpt_path, 'w+') as ckpt_file:
+            ckpt_file.write(str(chunk_idx))
+
     sf.set_semantic_constraints()  # restore defaults
-
-    assert not error_found_flag
-
-
-# TODO: Comment out the pytest skip to use this script.
-#  This test is somewhat covered by the first test though.
-@pytest.mark.skip(reason="eMolecules dataset not on GitHub")
-@pytest.mark.parametrize("test_name", datasets)
-def test_kekulize_parser(test_name):
-    """Tests the kekulization of SMILES, which is the first step of
-    selfies.encoder().
-    """
-
-    # file I/O
-    curr_dir = os.path.dirname(__file__)
-    test_path = os.path.join(curr_dir, 'test_sets', 'eMolecules',
-                             f"{test_name}.txt")
-    error_path = os.path.join(curr_dir,
-                              'error_sets',
-                              f"errors_kekulize_{test_name}.csv")
-
-    os.makedirs(os.path.dirname(error_path), exist_ok=True)
-    error_list = []
-    with open(error_path, "w+") as error_log:
-        error_log.write("In\n")
-    error_found_flag = False
-
-    # make pandas reader
-    N = sum(1 for _ in open(test_path)) - 1
-    S = NUM_SAMPLES if (0 < NUM_SAMPLES <= N) else N
-    skip = sorted(random.sample(range(1, N + 1), N - S))
-    reader = pd.read_csv(test_path,
-                         chunksize=10000,
-                         header=0,
-                         skiprows=skip)
-
-    # kekulize testing
-    for chunk in reader:
-        for smiles in chunk[COL_NAME]:
-
-            if (MolFromSmiles(smiles) is None) or ('*' in smiles):
-                continue
-
-            # build kekulized SMILES
-            kekule_fragments = []
-
-            for fragment in smiles.split("."):
-
-                kekule_gen = kekulize_parser(_parse_smiles(fragment))
-
-                k = []
-                for bond, symbol, symbol_type in kekule_gen:
-                    if symbol_type == BRANCH_TYPE:
-                        bond = ''
-                    k.append(bond)
-
-                    if symbol_type == RING_TYPE and len(symbol) == 2:
-                        k.append('%')
-                    k.append(symbol)
-
-                kekule_fragments.append(''.join(k))
-
-            kekule_smiles = '.'.join(kekule_fragments)
-
-            if not is_same_mol(smiles, kekule_smiles):
-                error_list.append(smiles)
-
-        with open(error_path, "a") as error_log:
-            error_log.write("\n".join(error_list))
-        error_found_flag = error_found_flag or error_list
-        error_list = []
+    os.remove(ckpt_path)  # remove checkpoint
 
     assert not error_found_flag
